diploma/requirements.txt
Django==2.1.1
django-user-accounts==2.1.0
django-appconf==1.0.3

